refactory SegmentDestination.kt
